// src/pages/Decorator/test.d.ts
  // global.d.ts
//   declare module 'global_type' {
//     export type IAnimal = {
//       name: string;
//     };
//     export type ICat = {
//       name: string;
//     };
//   }
//   declare module 'global_type1' {
//     export type IDog = {
//       name: string;
//     };
//   }
 

export type IAnimal = {
    name: string
    color?: string
}
export type ICat = {
    name: string
    color?: string
}
export type IDog = {
    name: string
    color?: string
}